restructuring, debugging
