+++ scripts/MIRfix.py
@@ -9,7 +9,9 @@ import sys
 import re
 import inspect
 import shlex
+import time
 import multiprocessing
+from multiprocessing import set_start_method
 import numpy as np
 import matplotlib
 matplotlib.use('Agg')
@@ -19,24 +21,21 @@ import json
 import gzip
 import traceback as tb
 from distutils.spawn import find_executable
-
-##import own modules
-from lib.Collection import *
-from lib.logger import makelogdir, setup_multiprocess_logger
-# Create log dir
-makelogdir('logs')
-# Define loggers
-global scriptname, streamlog, log           # global to ensure that later manipulation of loglevel is possible
-scriptname=os.path.basename(__file__)
-
 ##import Bio modules
 from Bio import SeqIO
 from Bio.Seq import Seq
 from Bio import AlignIO
 from Bio.Align.Applications import ClustalwCommandline
-
 ##import ViennaRNA
 import RNA
+#logging
+import datetime
+from lib.logger import makelogdir, setup_multiprocess_logger
+# Create log dir
+makelogdir('LOGS')
+scriptname=os.path.basename(__file__)
+##import own modules
+from lib.Collection import *
 
 def getindex(sequence,specie,precID,precdesc,listnogenomes,listnotingenome,templong):#get the index of the original sequence in its genome
     logid = scriptname+'.getindex: '
@@ -358,30 +357,6 @@ def flip(filename,filen,outdir,mappingfile,matfile,listofnew,listofnewloop,listo
         log.error(logid+''.join(tbe.format()))
 
 
-def dofold(listnewold,oldid,precseq,newid,newseq):
-    logid = scriptname+'.dofold: '
-    try:
-        md = RNA.md()
-        md.dangles = 2 #int(sys.argv[10]) if sys.argv[10] else 3   JF: DO NOT CHANGE
-        md.noLP=1
-        fc = RNA.fold_compound(str(precseq), md)
-        (oldstruct,oldscore) = fc.mfe()
-        mdn = RNA.md()
-        mdn.dangles = 2 #int(sys.argv[10]) if sys.argv[10] else 3   JF: DO NOT CHANGE
-        mdn.noLP=1
-        fcn = RNA.fold_compound(str(newseq), mdn)
-        (newstruct,newscore)= fcn.mfe()
-        log.debug(["folding is here:",oldstruct,oldscore])
-        log.debug(["folding is here:",newstruct,newscore])
-        listnewold.extend((oldid,str(precseq),oldstruct,oldscore,newid,str(newseq),newstruct,newscore))
-        return listnewold
-
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        log.error(logid+''.join(tbe.format()))
 
 
 def readfold(listnewold,filename,oldlstlstr,oldlstlstl,spos,epos,newspos,newepos,matdesc,matseq,outdir,oldparts,finaloldcomp,precDes,listofnew,listofnewloop,listoldstatus,listofoldloop,listofold,listofboth,listofmirstar,listnomat,listgoodnew):
@@ -2049,38 +2024,6 @@ def readfold(listnewold,filename,oldlstlstr,oldlstlstl,spos,epos,newspos,newepos
         )
         log.error(logid+''.join(tbe.format()))
 
-
-def comp_5p(ll, lr, sm, ss, precursor, run):
-    logid = scriptname+'.comp_5p: '
-    try:
-        log.debug(['comp5p', ll, lr, sm, ss, precursor, run])
-        if run > 3:
-            return ss-run+1
-        if sm in ll:
-            if any(x in precursor[sm].upper() for x in ['G', 'C']):
-                return ss-run+1
-            if ss in lr :
-                if any(x in precursor[sm].upper() for x in ['A', 'U']) and any(x in precursor[ss].upper() for x in ['G', 'C']):
-                    return ss
-                else:
-                    return comp_5p(ll, lr, sm, ss+1, precursor, run+1)
-            else:
-                return comp_5p(ll, lr, sm, ss+1, precursor, run+1)
-        else:
-            if ss in lr:
-                return ss
-            else:
-                return comp_5p(ll, lr, sm, ss+1, precursor, run+1)
-
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        log.error(logid+''.join(tbe.format()))
-
-
-
 def getmirstar(spos,epos,mature,lstl,lstr,precursor,hairpstart,hairpend):
     logid = scriptname+'.getmirstar: '
     mirstarepos=0 #CAVH
@@ -2348,89 +2291,6 @@ def getmirstarbak(spos,epos,mature,lstl,lstr,precursor,hairpstart,hairpend):
         )
         log.error(logid+''.join(tbe.format()))
 
-
-def foldnomat(inputfasta,outputfasta):#fold the new and the old sequences, using temp file every time I get the new sequence from the original
-    logid = scriptname+'.foldnomat: '
-    try:
-        f=os.popen("RNAfold -d3 --noPS --noLP <"+inputfasta)
-        fi=f.read()
-        wr=open(outputfasta,"w")
-        f.close()
-        wr.write(fi)
-        wr.close
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        log.error(logid+''.join(tbe.format()))
-
-
-def alignTostock(align):
-    logid = scriptname+'.alignTostock: '
-    try:
-        reads=openfile(align)
-        stkfile=align+".stk"
-        writes=open(stkfile,'a')
-        numberlines=0#count lines
-        items=[]
-        listofids=[]
-        maxspaces=0
-        minlen=100000
-        maxlen=0
-        for line in reads:
-            item=line.split()
-            if len(item) > 0 and item[0]!="CLUSTAL" and ("*" not in line):
-                numberlines=numberlines+1
-                listofids.append(item[0])
-                items.append(line.strip())
-                numberofspaces=items[0].count(" ")
-                if maxspaces<numberofspaces:
-                    maxspaces=numberofspaces
-        for i in listofids:
-            s=len(i)
-            if minlen>s:
-                minlen=s
-            if maxlen<s:
-                maxlen=s
-        if maxlen>len('#=GC SS_cons'):
-            maxlen=maxlen
-        if maxlen<=len('#=GC SS_cons'):
-            maxlen=len('#=GC SS_cons')
-
-        numberofblocks=listofids.count(listofids[0])
-        numberofseqs=int(numberlines/numberofblocks)
-        ranges=int(len(items)/2)
-        writes.write('# STOCKHOLM 1.0'+"\n"+"\n")
-        offset=0
-        for i in range (0,numberofseqs):
-            offset=0
-
-            seqitem=items[i].split()
-            seqid=seqitem[0]
-            seq=seqid.strip()+" "*(maxlen-len(seqid)+1)+seqitem[1]
-
-            for k in range(0,numberofblocks-1):
-                offset=offset+numberofseqs
-                item2=items[i+offset].split()
-                seq=seq+item2[1]
-
-            writes.write(seq+"\n")
-        tempitem=seq.split()
-        writes.write('#=GC SS_cons'+" "*(maxlen-len('#=GC SS_cons')+1)+"."*len(tempitem[1])+"\n"+"//")
-        writest=str(writes)
-        writeitem=writest.split("'")
-        writes.close()
-        return str(writeitem[1])
-
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        log.error(logid+''.join(tbe.format()))
-
-
 def predict(align,matId,newmatID,matfile,filename,precdescrip,mapfile,directory,listremovedbroken,listremovedscore):
     logid = scriptname+'.predict: '
     try:
@@ -2653,27 +2513,6 @@ def checknomat(precfile,mapfile,matfile,directory,precfilename,listremovedbroken
             flagnomatexists=False
             log.debug("all precursors has mature, at least one")
         return flagnomatexists,nomats,listremovedbroken,listremovedscore,listnomat
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        log.error(logid+''.join(tbe.format()))
-
-
-def getfilename(dirfile):#to get the name of the file without directory or extension, in principal was used to get famname
-    logid = scriptname+'.getfilename: '
-    try:
-        if '/' in dirfile and "." in dirfile:
-            dirf=dirfile[dirfile.rfind('/')+1:]
-            filename=dirf[:dirf.rfind('.')]
-        elif '/' in dirfile and "." not in dirfile:
-            filename=dirfile[dirfile.rfind('/')+1:]
-        elif '/' not in dirfile and "." in dirfile:
-            filename=dirfile[:dirfile.rfind('.')]
-        elif '/' not in dirfile and "." not in dirfile:
-            filename=dirfile
-        return str(filename)
     except Exception as err:
         exc_type, exc_value, exc_tb = sys.exc_info()
         tbe = tb.TracebackException(
@@ -3007,169 +2846,6 @@ def readfoldpredict(foldfile,predictedspos,predictedepos):#,predictedspos,predic
         )
         log.error(logid+''.join(tbe.format()))
 
-
-def doalifold(alnfile,outdir):
-    logid = scriptname+'.doalifold: '
-    try:
-        f=os.popen("RNAalifold --noPS -r "+alnfile) # -r = RIBOSUM scoring
-        alifoldtemp=outdir+'alifoldtemp.txt'
-        foldrestemp=open(alifoldtemp,'w')
-        fi=f.read()
-        foldrestemp.write(fi)
-        foldrestemp.close()
-        f.close()
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        log.error(logid+''.join(tbe.format()))
-
-
-def getstructure(alifoldtemp):
-    logid = scriptname+'.getstructure: '
-    try:
-        for line in (openfile(alifoldtemp)):
-            item=line.split()
-            hairpin = ''
-            if len(item)>0 and 'A' not in item[0] and 'C' not in item[0] and 'G' not in item[0] and 'T' not in item[0] and 'a' not in item[0] and 'c' not in item[0] and 'g' not in item[0] and 't' not in item[0] and 'N' not in item[0] and 'n' not in item[0] and ('.' in item[0] or '(' in item[0] or ')' in item[0]):
-                hairpin=item[0]
-                return str(hairpin)
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        log.error(logid+''.join(tbe.format()))
-
-
-def cal_ent(s):
-    logid = scriptname+'.cal_ent: '
-    try:
-        listofchars=[0,0,0,0,0]
-
-        seqi=s
-        lengseqi=len(seqi)
-        listofsites = []
-        for i in seqi:#loop on the sequences on each site
-            if i == 'A'or i =='a':
-                listofchars[0]=(listofchars[0]+1)#/numofseqs
-            if i == 'C'or i =='c':
-                listofchars[1]=(listofchars[1]+1)#/numofseqs
-            if i == 'G'or i =='g':
-                listofchars[2]=(listofchars[2]+1)#/numofseqs
-            if i == 'T'or i =='t' or i =='u' or i =='U'  :
-                listofchars[3]=(listofchars[3]+1)#/numofseqs
-            if i == '-'or i =='.':
-                listofchars[4]=(listofchars[4]+1)#/numofseqs
-
-        listofsites.append(listofchars)#list of sites adding the ratio for each of them
-        listofchars = [0,0,0,0,0]
-        for k in range (0,5):#find ratio of each nucleotide at a specific site
-            listofchars[k]=listofchars[k]/lengseqi
-
-        listofsites.append(listofchars)#list of sites adding the ratio for each of them
-        PA=0
-        PC=0
-        PG=0
-        PT=0
-        PGA=0
-        PA=(listofsites[0][0])/lengseqi
-        PC=(listofsites[0][1])/lengseqi
-        PG=(listofsites[0][2])/lengseqi
-        PT=(listofsites[0][3])/lengseqi
-        PGA=(listofsites[0][4])/lengseqi
-        EA=0
-        EC=0
-        EG=0
-        ET=0
-        EGA=0
-
-        if PA > 0:
-            EA=(PA*(math.log(PA,2)))
-        else:
-            EA=0
-
-        if PC > 0:
-            EC=(PC*(math.log(PC,2)))
-        else:
-            EC=0
-
-        if PG > 0:
-            EG=(PG*(math.log(PG,2)))
-        else:
-            EG=0
-
-        if PT > 0:
-            ET=(PT*(math.log(PT,2)))
-        else:
-            ET=0
-
-        if PGA > 0:
-            EGA= (PGA*(math.log(PGA,2)))
-        else:
-            EGA=0
-
-        #Entropy=-1* ((PA*(math.log(PA,2)))+ (PC*(math.log(PC,2))) + (PG*(math.log(PG,2))) + (PT*(math.log(PT,2))) + (PGA*(math.log(PGA,2))))
-        Entropy=-1* (EA+EC+EG+ET+EGA)
-
-        return Entropy
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        log.error(logid+''.join(tbe.format()))
-
-
-def CalShanon(stkfile):
-    logid = scriptname+'.CalShanon: '
-    try:
-        aligEnt=0
-        nameOffile=stkfile
-        listofchars = [] #use as temp to count nucleotides at each site
-        listofsites = [] #use to store the ratio of each nucleotide at each site, i filled it as list of list
-        lengseq=0  #lenght of sequence, NOTE: all should be same
-        numofseqs=0
-        #loop to initialize nucleotides' list to zero
-        for i in range (0,5):
-            listofchars.append(0)
-
-            #Fill the list of sites with the ratios of nucleotides
-        alignment = AlignIO.read(nameOffile, "stockholm")
-
-        for record1 in  alignment:
-            numofseqs+=1
-            recid1= record1.id
-            seqi=str(record1.seq)
-            lengseq=len(seqi)
-
-        SeqEachSite=[]
-        EntEachSite=[]
-
-        for i in  range (0,lengseq):
-            EntEachSite.append("")
-            SeqEachSite.append("")
-
-        for record in alignment:
-            seqtemp=str(record.seq)
-            for i in range(0,lengseq):
-                SeqEachSite[i]=SeqEachSite[i]+seqtemp[i]
-
-        for i in range(0,lengseq):
-            EntEachSite=cal_ent(SeqEachSite[i])
-            aligEnt=aligEnt+EntEachSite
-
-        return aligEnt
-
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        log.error(logid+''.join(tbe.format()))
-
-
 def correct(corid,flanking,countcorrected,countcorrectedTonew,listofnew,listofnewloop,listoldstatus,templong,listmisalignedcorr,listcorrected,listcorrectedori,listgoodnew):
     logid = scriptname+'.correct: '
     try:
@@ -3334,7 +3010,10 @@ def correct(corid,flanking,countcorrected,countcorrectedTonew,listofnew,listofne
         log.error(logid+''.join(tbe.format()))
 
 def sublist(filename):
+    print('HERE')
     logid = scriptname+'.sublist: '
+    log.debug(logid+'Starting to process '+str(filename))
+    print('HERE')
     try:
         filesdir=str(args.famdir)#dir for families
         command="list"
@@ -4717,19 +4396,6 @@ def sublist(filename):
         )
         log.error(logid+''.join(tbe.format()))
 
-
-def openfile(f):
-    logid = scriptname+'.openfile: '
-    try:
-        return open(f,'r') if not '.gz' in f[-4:] else gzip.open(f,'rt')
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        log.error(logid+''.join(tbe.format()))
-
-
 def parseargs():
     parser = argparse.ArgumentParser(description='MIRfix automatically curates miRNA datasets by improving alignments of their precursors, the consistency of the annotation of mature miR and miR* sequence, and the phylogenetic coverage. MIRfix produces alignments that are comparable across families and sets the stage for improved homology search as well as quantitative analyses.')
     parser.add_argument("-j", "--cores", type=int, default=1, help='Number of parallel processes to run')
@@ -4749,188 +4415,6 @@ def parseargs():
 
     return parser.parse_args()
 
-### CAVH functions
-
-def count_repetitions(longseq,pattern):
-    results=0
-    len_pattern = len(pattern)
-    for i in range(len(longseq)):
-        if longseq[i:i+len_pattern] == pattern:
-            results += 1
-            print(longseq[i:i+len_pattern],i,i+len_pattern)
-    return results
-
-
-def search_pattern(longseq, pattern):
-    start = longseq.find(pattern)
-    end = int((start + len(pattern)) -1)
-    return (start, end)
-
-
-def generate_array(longseq, pattern, number, mode):
-    pattern_len = len(pattern)
-    array = []
-    short_sequence = None
-    update_index = None
-    for j in range(0, number):  # (0, n-1)
-        if j == 0:  # Search on the complete sequence
-            start = longseq.find(pattern)
-            end = int((start + pattern_len) - 1)
-            assembly_data = [start, end, mode]
-            array.append(assembly_data)
-            short_sequence = longseq[end + 1:]
-            update_index = longseq.find(short_sequence)
-        else:
-            start_temp = short_sequence.find(pattern)
-            if start_temp != -1:
-                end_temp = int((start_temp + pattern_len) - 1)
-                start = int(start_temp + update_index)
-                end = int(end_temp + update_index)
-                assembly_data = [start, end, mode]
-                array.append(assembly_data)
-                short_sequence = short_sequence[end_temp + 1:]
-                update_index = longseq.find(short_sequence)
-    return array
-
-
-def detect_overlapp_in_array(complete_array):
-    columns = len(complete_array)
-    for i in range(0, columns -1):
-        if complete_array[i+1][0] in range(complete_array[i][0],
-           complete_array[i][1] + 1):
-            return 1
-    return 0
-
-
-def detect_pairs(complete_array, longseq, mat1seq, curmatseq, curmatstar):
-    columns = len(complete_array)  # Complete number columns
-    # Detect overlapping
-    overlapping = detect_overlapp_in_array(complete_array)
-    pair = defaultdict(dict)
-    number_pair = 0
-    if overlapping == 0:
-        for i in range(0, columns-1):  # In array =columns -1,range not reach end#
-            if complete_array[i][-1] != complete_array[i+1][-1]:
-                pair[number_pair] = [complete_array[i], complete_array[i+1]]
-                number_pair = number_pair + 1
-    else:  # At least one pair overlapped with another one
-        substring = None
-        update_index = None
-        for j in range(0, columns -1):  # (0, n-1)
-            if j == 0:  # Search on the complete sequence
-                identity_first = complete_array[j][-1]  # M or S
-                if identity_first == "M":
-                    start_m = longseq.find(curmatseq)  # Start M
-                    end_m = (start_m + len(curmatseq)) - 1  # End M
-                    substring = longseq[end_m + 1:]  # Restrict seq no M
-                    start_temp_s = substring.find(curmatstar)  # Detect start S
-                    update_index = longseq.find(substring)  # Get index
-                    start_s = start_temp_s + update_index  # Update index S
-                    end_s = (start_s + len(curmatstar)) -1
-                    element1 = [start_m, end_m, "M"]
-                    element2 = [start_s, end_s, "S"]
-                    pair[number_pair] = [element1, element2]
-                    number_pair = number_pair + 1
-                else:  # When S is first, search M
-                    start_s = longseq.find(curmatstar)  # Start S
-                    end_s = (start_s + len(curmatstar))-1  # End S
-                    substring = longseq[end_s + 1:]  # Restrict seq no S
-                    update_index = longseq.find(substring)
-                    start_temp_m = substring.find(curmatseq)  # Search M
-                    start_m = start_temp_m + update_index  # Start M
-                    end_m = (start_m + len(curmatseq)) - 1  # End M
-                    element1 = [start_s, end_s, "S"]
-                    element2 = [start_m, end_m, "M"]
-                    pair[number_pair] = [element1, element2]
-                    number_pair = number_pair + 1
-            else:  # This is not the first element
-                if complete_array[j][0] in range(complete_array[j-1][0],
-                                                 complete_array[j-1][1]):
-                    continue
-                else:
-                    identity_first = complete_array[j][-1]  # M or S
-                    if identity_first == "M":
-                        start_temp_m = substring.find(curmatseq)  # Start M in sub
-                        end_temp_m = (start_temp_m + len(curmatseq)) - 1
-                        start_m = start_temp_m + update_index  # Updated M Start
-                        end_m = (start_m + len(curmatseq)) - 1  # Updated M end
-                        substring = substring[end_temp_m + 1:]   # String no start M
-                        update_index = longseq.find(substring)
-                        start_temp_s = substring.find(curmatstar)
-                        if (start_temp_s == -1):  # Not possible to map S
-                            continue
-                        start_s = start_temp_s + update_index  # Updated start S
-                        end_s = (start_s + len(curmatstar)) -1  # Updated end S
-                        element1 = [start_m, end_m, "M"]
-                        element2 = [start_s, end_s, "S"]
-                        pair[number_pair] = [element1, element2]
-                        number_pair = number_pair + 1
-                    else:  # When S is first, search M
-                        start_temp_s = substring.find(curmatstar)  # Start M in sub
-                        end_temp_s = (start_temp_s + len(curmatstar)) - 1
-                        start_s = start_temp_s + update_index  # Updated S Start
-                        end_s = (start_s + len(curmatstar)) - 1  # Updated S end
-                        substring = substring[end_temp_s + 1:]   # String no start S
-                        update_index = longseq.find(substring)
-                        start_temp_m = substring.find(curmatseq)
-                        start_m = start_temp_m + update_index  # Update start M
-                        if (start_temp_m == -1):  # Not possible to map M
-                            continue
-                        end_m = (start_s + len(curmatseq)) -1  # Updated end M
-                        element1 = [start_s, end_s, "S"]
-                        element2 = [start_m, end_m, "M"]
-                        pair[number_pair] = [element1, element2]
-                        number_pair = number_pair + 1
-    return (pair, number_pair)
-
-
-def define_best_pair(pairs, distances_loop, startprecursor,
-                     endprecursor, precursorlen, flanking, longseq):
-    all_distances_sorted = sorted(distances_loop, key=itemgetter(-1))
-    for i in range(len(all_distances_sorted)):
-        if (all_distances_sorted[i][1][0][0] in range(startprecursor, endprecursor + 1)
-                and all_distances_sorted[i][1][0][1] in range(startprecursor, endprecursor + 1)
-                and all_distances_sorted[i][1][1][0] in range(startprecursor, endprecursor + 1)
-                and all_distances_sorted[i][1][1][1] in range(startprecursor, endprecursor + 1)):
-            coord1 = all_distances_sorted[i][1][0][0]
-            coord2 = all_distances_sorted[i][1][1][0]
-            finalcoord2 = all_distances_sorted[i][1][1][1]
-            startfinalseq = coord1-flanking
-            endfinalseq = finalcoord2+flanking
-            correctfinalseq = longseq[startfinalseq:endfinalseq+1]
-            break
-    return (coord1, coord2, correctfinalseq)
-
-
-def find_positions(longseq, mat1seq, curmatseq, curmatstar, flanking):
-    # Features sequences
-    precursorlen = len(mat1seq)
-    # Locate
-    startprecursor = longseq.find(mat1seq)
-    endprecursor = int((startprecursor + precursorlen)-1)
-    # Search the number of substrings along the long sequence
-    mir = count_repetitions(longseq,curmatseq)
-    mirstar = count_repetitions(longseq,curmatstar)
-    mir_array = generate_array(longseq, curmatseq, mir, "M")
-    mirstar_array = generate_array(longseq, curmatstar, mirstar, "S")
-    # All array
-    all_array = mir_array + mirstar_array
-    all_array_sorted = sorted(all_array, key=itemgetter(0))
-    (pairs, number_pairs) = detect_pairs(all_array_sorted, longseq,
-                                         mat1seq, curmatseq, curmatstar)
-    distances_loop = []
-    for pair in pairs:
-        loopsize = abs(pairs[pair][0][1] - pairs[pair][1][0]) - 1
-        distances_loop.append([pair, pairs[pair], loopsize])
-    (coord1, coord2, correctfinalseq) = define_best_pair(pairs, distances_loop,
-                                                         startprecursor,
-                                                         endprecursor,
-                                                         precursorlen,
-                                                         flanking,
-                                                         longseq)
-    return (coord1, coord2, correctfinalseq)
-
-
 ##############################MAIN##############################
 
 if __name__ == '__main__':
@@ -4938,15 +4422,30 @@ if __name__ == '__main__':
     logid = scriptname+'.main: '
 
     try:
-        global args
-        args = parseargs()
-        log = setup_multiprocess_logger(name=logid, log_file='logs/'+scriptname, logformat='%(asctime)s %(name)-12s %(levelname)-8s %(message)s', datefmt='%m-%d %H:%M', level=args.loglevel)
-        streamlog = setup_multiprocess_logger(name='', log_file='stderr', logformat='%(asctime)s %(name)-12s %(levelname)-8s %(message)s', datefmt='%m-%d %H:%M', level=args.loglevel)
+        set_start_method("spawn")  # set multiprocessing start method to safe spawn
+        args=parseargs()
+        logtime = str(datetime.datetime.now().strftime("%Y%m%d_%H_%M_%S_%f"))
+        logdir = 'LOGS'+os.sep+logtime
+        makelogdir(logdir)
+        logfile = os.path.abspath(logdir)+os.sep+scriptname+'.log'
 
+        #find_executable('clustalw2') or sys.exit('Please install clustalw2 to run this')
+        find_executable('dialign2-2') or sys.exit('Please install dialign2-2 to run this')
+
+        if not os.path.isfile(os.path.abspath(logfile)):
+            open(logfile,'a').close()
+        else:
+            ts = str(datetime.datetime.fromtimestamp(os.path.getmtime(os.path.abspath(logfile))).strftime("%Y%m%d_%H_%M_%S"))
+            shutil.move(logfile,logdir+os.sep+scriptname+'_'+ts+'.log')
+
+        log = setup_multiprocess_logger(log_file='stderr', logformat='%(asctime)s %(levelname)-8s %(name)-12s %(message)s', datefmt='%m-%d %H:%M')
+        log = setup_multiprocess_logger(log_file=logfile, filemode='a', logformat='%(asctime)s %(levelname)-8s %(name)-12s %(message)s', datefmt='%m-%d %H:%M')
+
+        log.setLevel(args.loglevel)
         log.info(logid+'Running '+scriptname+' on '+str(args.cores)+' cores.')
-        log.info(logid+'CLI: '+str(sys.argv[0]) + ' '.join( [shlex.quote(s) for s in sys.argv[1:]] ))
+        log.info(logid+'CLI: '+sys.argv[0]+' '+'{}'.format(' '.join( [shlex.quote(s) for s in sys.argv[1:]] )))
 
-        nthreads=args.cores
+        nthreads=args.cores or 2
         lfams = []
 
         with openfile(args.families) as filelist:
@@ -4959,12 +4458,9 @@ if __name__ == '__main__':
 
         pool = multiprocessing.Pool(processes=nthreads, maxtasksperchild=1)
 
-        #find_executable('clustalw2') or sys.exit('Please install clustalw2 to run this')
-        find_executable('dialign2-2') or sys.exit('Please install dialign2-2 to run this')
-
         for fam in lfams:
-            pool.apply_async(sublist, args=(fam,))
-
+            pool.apply_async(sublist, args=(fam))
+            log.info('Working on '+str(fam))
         pool.close()
         pool.join()
