+++ scripts/lib/Collection.py
@@ -1,77 +1,5 @@
-### Collection.py ---
-##
-## Filename: Collection.py
-## Description:
-## Author: Joerg Fallmann
-## Maintainer:
-## Created: Thu Sep  6 09:02:18 2018 (+0200)
-## Version:
-## Package-Requires: ()
-## Last-Updated: Tue Nov  5 15:56:40 2019 (+0100)
-##           By: Joerg Fallmann
-##     Update #: 252
-## URL:
-## Doc URL:
-## Keywords:
-## Compatibility:
-##
-######################################################################
-##
-### Commentary:
-###import os, sys, inspect
-# # realpath() will make your script run, even if you symlink it :)
-# cmd_folder = os.path.dirname(os.path.realpath(os.path.abspath( inspect.getfile( inspect.currentframe() )) ))
-# if cmd_folder not in sys.path:
-#     sys.path.insert(0, cmd_folder)
-#
-# # Use this if you want to include modules from a subfolder
-# cmd_subfolder = os.path.join(os.path.dirname(os.path.realpath(os.path.abspath( inspect.getfile( inspect.currentframe() )) )),"lib")
-# if cmd_subfolder not in sys.path:
-#     sys.path.insert(0, cmd_subfolder)
-#
-# # Info:
-# # cmd_folder = os.path.dirname(os.path.abspath(__file__)) # DO NOT USE __file__ !!!
-# # __file__ fails if the script is called in different ways on Windows.
-# # __file__ fails if someone does os.chdir() before.
-# # sys.argv[0] also fails, because it doesn't not always contains the path.
-##
-##
-######################################################################
-##
-### Change Log:
-##
-##
-######################################################################
-##
-## This program is free software: you can redistribute it and/or modify
-## it under the terms of the GNU General Public License as published by
-## the Free Software Foundation, either version 3 of the License, or (at
-## your option) any later version.
-##
-## This program is distributed in the hope that it will be useful, but
-## WITHOUT ANY WARRANTY; without even the implied warranty of
-## MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-## General Public License for more details.
-##
-## You should have received a copy of the GNU General Public License
-## along with GNU Emacs.  If not, see <http://www.gnu.org/licenses/>.
-##
-######################################################################
-##
-### Code:
-### IMPORTS
 import os, sys, inspect
-
-##load own modules
-from lib.logger import makelogdir, setup_multiprocess_logger
-# Create log dir
-makelogdir('logs')
-# Define loggers
-scriptname=os.path.basename(__file__)
-#streamlog = setup_multiprocess_logger(name='', log_file='stderr', logformat='%(asctime)s %(name)-12s %(levelname)-8s %(message)s', datefmt='%m-%d %H:%M', level='WARNING')
-clog = setup_multiprocess_logger(name=scriptname, log_file='logs/'+scriptname, logformat='%(asctime)s %(name)-12s %(levelname)-8s %(message)s', datefmt='%m-%d %H:%M', level='WARNING')
-from lib.Collection import *
-
+from lib.logger import *
 ##other modules
 import numpy as np
 import heapq
@@ -86,63 +14,525 @@ import os
 import gzip
 import math
 from collections import defaultdict
-#Biopython stuff
+##import Bio modules
 from Bio import SeqIO
 from Bio.Seq import Seq
+from Bio import AlignIO
+from Bio.Align.Applications import ClustalwCommandline
+
+scriptname = os.path.basename(inspect.stack()[-1].filename).replace('.py','')
 
 # Code:All subs from here on
-def create_kmers(choices, length):
-    logid = scriptname+'.create_kmers: '
+#Files
+def openfile(f):
+    logid = scriptname+'.openfile: '
     try:
-        #     choices=['A','T','G','C']
-        bases = list(choices)
-        k = length
+        return open(f,'r') if not '.gz' in f[-4:] else gzip.open(f,'rt')
+    except Exception as err:
+        exc_type, exc_value, exc_tb = sys.exc_info()
+        tbe = tb.TracebackException(
+            exc_type, exc_value, exc_tb,
+        )
+        log.error(logid+''.join(tbe.format()))
 
-        return [''.join(p) for p in itertools.product(bases, repeat=k)]
+def getfilename(dirfile):#to get the name of the file without directory or extension, in principal was used to get famname
+    logid = scriptname+'.getfilename: '
+    try:
+        if '/' in dirfile and "." in dirfile:
+            dirf=dirfile[dirfile.rfind('/')+1:]
+            filename=dirf[:dirf.rfind('.')]
+        elif '/' in dirfile and "." not in dirfile:
+            filename=dirfile[dirfile.rfind('/')+1:]
+        elif '/' not in dirfile and "." in dirfile:
+            filename=dirfile[:dirfile.rfind('.')]
+        elif '/' not in dirfile and "." not in dirfile:
+            filename=dirfile
+        return str(filename)
+    except Exception as err:
+        exc_type, exc_value, exc_tb = sys.exc_info()
+        tbe = tb.TracebackException(
+            exc_type, exc_value, exc_tb,
+        )
+        log.error(logid+''.join(tbe.format()))
+
+#foldings
+def dofold(listnewold,oldid,precseq,newid,newseq):
+    logid = scriptname+'.dofold: '
+    try:
+        md = RNA.md()
+        md.dangles = 2 #int(sys.argv[10]) if sys.argv[10] else 3   JF: DO NOT CHANGE
+        md.noLP=1
+        fc = RNA.fold_compound(str(precseq), md)
+        (oldstruct,oldscore) = fc.mfe()
+        mdn = RNA.md()
+        mdn.dangles = 2 #int(sys.argv[10]) if sys.argv[10] else 3   JF: DO NOT CHANGE
+        mdn.noLP=1
+        fcn = RNA.fold_compound(str(newseq), mdn)
+        (newstruct,newscore)= fcn.mfe()
+        log.debug(["folding old is here:",oldstruct,oldscore])
+        log.debug(["folding new is here:",newstruct,newscore])
+        listnewold.extend((oldid,str(precseq),oldstruct,oldscore,newid,str(newseq),newstruct,newscore))
+        return listnewold
 
     except Exception as err:
         exc_type, exc_value, exc_tb = sys.exc_info()
         tbe = tb.TracebackException(
             exc_type, exc_value, exc_tb,
         )
-        clog.error(logid+''.join(tbe.format()))
+        log.error(logid+''.join(tbe.format()))
 
-def randseq(alphabet, length):
-    logid = scriptname+'.randseq: '
+def foldnomat(inputfasta,outputfasta):#fold the new and the old sequences, using temp file every time I get the new sequence from the original
+    logid = scriptname+'.foldnomat: '
     try:
-        l=''
-        for i in range(length):
-            l += str(''.join(choice(alphabet)))
-            return str(l)
+        f=os.popen("RNAfold -d3 --noPS --noLP <"+inputfasta)
+        fi=f.read()
+        wr=open(outputfasta,"w")
+        f.close()
+        wr.write(fi)
+        wr.close
     except Exception as err:
         exc_type, exc_value, exc_tb = sys.exc_info()
         tbe = tb.TracebackException(
             exc_type, exc_value, exc_tb,
         )
-        clog.error(logid+''.join(tbe.format()))
+        log.error(logid+''.join(tbe.format()))
 
-def weightedrandseq(alphabet, probs , length):
-    logid = scriptname+'.weightedrandseq: '
+def doalifold(alnfile,outdir):
+    logid = scriptname+'.doalifold: '
     try:
-        seq=[]
-        for i in alphabet:
-            weight=int(next(probs))
-            for a in range(weight):
-                seq+=i
-
-        if len(seq) < length:
-            for i in range(length-len(seq)):
-                seq += str(''.join(choice(alphabet)))
-
-        shuffle(seq)
-        return seq
+        f=os.popen("RNAalifold --noPS -r "+alnfile) # -r = RIBOSUM scoring
+        alifoldtemp=outdir+'alifoldtemp.txt'
+        foldrestemp=open(alifoldtemp,'w')
+        fi=f.read()
+        foldrestemp.write(fi)
+        foldrestemp.close()
+        f.close()
     except Exception as err:
         exc_type, exc_value, exc_tb = sys.exc_info()
         tbe = tb.TracebackException(
             exc_type, exc_value, exc_tb,
         )
-        clog.error(logid+''.join(tbe.format()))
+        log.error(logid+''.join(tbe.format()))
 
+def getstructure(alifoldtemp):
+    logid = scriptname+'.getstructure: '
+    try:
+        for line in (openfile(alifoldtemp)):
+            item=line.split()
+            hairpin = ''
+            if len(item)>0 and 'A' not in item[0] and 'C' not in item[0] and 'G' not in item[0] and 'T' not in item[0] and 'a' not in item[0] and 'c' not in item[0] and 'g' not in item[0] and 't' not in item[0] and 'N' not in item[0] and 'n' not in item[0] and ('.' in item[0] or '(' in item[0] or ')' in item[0]):
+                hairpin=item[0]
+                return str(hairpin)
+    except Exception as err:
+        exc_type, exc_value, exc_tb = sys.exc_info()
+        tbe = tb.TracebackException(
+            exc_type, exc_value, exc_tb,
+        )
+        log.error(logid+''.join(tbe.format()))
+
+def cal_ent(s):
+    logid = scriptname+'.cal_ent: '
+    try:
+        listofchars=[0,0,0,0,0]
+
+        seqi=s
+        lengseqi=len(seqi)
+        listofsites = []
+        for i in seqi:#loop on the sequences on each site
+            if i == 'A'or i =='a':
+                listofchars[0]=(listofchars[0]+1)#/numofseqs
+            if i == 'C'or i =='c':
+                listofchars[1]=(listofchars[1]+1)#/numofseqs
+            if i == 'G'or i =='g':
+                listofchars[2]=(listofchars[2]+1)#/numofseqs
+            if i == 'T'or i =='t' or i =='u' or i =='U'  :
+                listofchars[3]=(listofchars[3]+1)#/numofseqs
+            if i == '-'or i =='.':
+                listofchars[4]=(listofchars[4]+1)#/numofseqs
+
+        listofsites.append(listofchars)#list of sites adding the ratio for each of them
+        listofchars = [0,0,0,0,0]
+        for k in range (0,5):#find ratio of each nucleotide at a specific site
+            listofchars[k]=listofchars[k]/lengseqi
+
+        listofsites.append(listofchars)#list of sites adding the ratio for each of them
+        PA=0
+        PC=0
+        PG=0
+        PT=0
+        PGA=0
+        PA=(listofsites[0][0])/lengseqi
+        PC=(listofsites[0][1])/lengseqi
+        PG=(listofsites[0][2])/lengseqi
+        PT=(listofsites[0][3])/lengseqi
+        PGA=(listofsites[0][4])/lengseqi
+        EA=0
+        EC=0
+        EG=0
+        ET=0
+        EGA=0
+
+        if PA > 0:
+            EA=(PA*(math.log(PA,2)))
+        else:
+            EA=0
+
+        if PC > 0:
+            EC=(PC*(math.log(PC,2)))
+        else:
+            EC=0
+
+        if PG > 0:
+            EG=(PG*(math.log(PG,2)))
+        else:
+            EG=0
+
+        if PT > 0:
+            ET=(PT*(math.log(PT,2)))
+        else:
+            ET=0
+
+        if PGA > 0:
+            EGA= (PGA*(math.log(PGA,2)))
+        else:
+            EGA=0
+
+        #Entropy=-1* ((PA*(math.log(PA,2)))+ (PC*(math.log(PC,2))) + (PG*(math.log(PG,2))) + (PT*(math.log(PT,2))) + (PGA*(math.log(PGA,2))))
+        Entropy=-1* (EA+EC+EG+ET+EGA)
+
+        return Entropy
+    except Exception as err:
+        exc_type, exc_value, exc_tb = sys.exc_info()
+        tbe = tb.TracebackException(
+            exc_type, exc_value, exc_tb,
+        )
+        log.error(logid+''.join(tbe.format()))
+
+def CalShanon(stkfile):
+    logid = scriptname+'.CalShanon: '
+    try:
+        aligEnt=0
+        nameOffile=stkfile
+        listofchars = [] #use as temp to count nucleotides at each site
+        listofsites = [] #use to store the ratio of each nucleotide at each site, i filled it as list of list
+        lengseq=0  #lenght of sequence, NOTE: all should be same
+        numofseqs=0
+        #loop to initialize nucleotides' list to zero
+        for i in range (0,5):
+            listofchars.append(0)
+
+            #Fill the list of sites with the ratios of nucleotides
+        alignment = AlignIO.read(nameOffile, "stockholm")
+
+        for record1 in  alignment:
+            numofseqs+=1
+            recid1= record1.id
+            seqi=str(record1.seq)
+            lengseq=len(seqi)
+
+        SeqEachSite=[]
+        EntEachSite=[]
+
+        for i in  range (0,lengseq):
+            EntEachSite.append("")
+            SeqEachSite.append("")
+
+        for record in alignment:
+            seqtemp=str(record.seq)
+            for i in range(0,lengseq):
+                SeqEachSite[i]=SeqEachSite[i]+seqtemp[i]
+
+        for i in range(0,lengseq):
+            EntEachSite=cal_ent(SeqEachSite[i])
+            aligEnt=aligEnt+EntEachSite
+
+        return aligEnt
+
+    except Exception as err:
+        exc_type, exc_value, exc_tb = sys.exc_info()
+        tbe = tb.TracebackException(
+            exc_type, exc_value, exc_tb,
+        )
+        log.error(logid+''.join(tbe.format()))
+
+#comparison
+def comp_5p(ll, lr, sm, ss, precursor, run):
+    logid = scriptname+'.comp_5p: '
+    try:
+        log.debug(['comp5p', ll, lr, sm, ss, precursor, run])
+        if run > 3:
+            return ss-run+1
+        if sm in ll:
+            if any(x in precursor[sm].upper() for x in ['G', 'C']):
+                return ss-run+1
+            if ss in lr :
+                if any(x in precursor[sm].upper() for x in ['A', 'U']) and any(x in precursor[ss].upper() for x in ['G', 'C']):
+                    return ss
+                else:
+                    return comp_5p(ll, lr, sm, ss+1, precursor, run+1)
+            else:
+                return comp_5p(ll, lr, sm, ss+1, precursor, run+1)
+        else:
+            if ss in lr:
+                return ss
+            else:
+                return comp_5p(ll, lr, sm, ss+1, precursor, run+1)
+
+    except Exception as err:
+        exc_type, exc_value, exc_tb = sys.exc_info()
+        tbe = tb.TracebackException(
+            exc_type, exc_value, exc_tb,
+        )
+        log.error(logid+''.join(tbe.format()))
+
+
+#format
+def alignTostock(align):
+    logid = scriptname+'.alignTostock: '
+    try:
+        reads=openfile(align)
+        stkfile=align+".stk"
+        writes=open(stkfile,'a')
+        numberlines=0#count lines
+        items=[]
+        listofids=[]
+        maxspaces=0
+        minlen=100000
+        maxlen=0
+        for line in reads:
+            item=line.split()
+            if len(item) > 0 and item[0]!="CLUSTAL" and ("*" not in line):
+                numberlines=numberlines+1
+                listofids.append(item[0])
+                items.append(line.strip())
+                numberofspaces=items[0].count(" ")
+                if maxspaces<numberofspaces:
+                    maxspaces=numberofspaces
+        for i in listofids:
+            s=len(i)
+            if minlen>s:
+                minlen=s
+            if maxlen<s:
+                maxlen=s
+        if maxlen>len('#=GC SS_cons'):
+            maxlen=maxlen
+        if maxlen<=len('#=GC SS_cons'):
+            maxlen=len('#=GC SS_cons')
+
+        numberofblocks=listofids.count(listofids[0])
+        numberofseqs=int(numberlines/numberofblocks)
+        ranges=int(len(items)/2)
+        writes.write('# STOCKHOLM 1.0'+"\n"+"\n")
+        offset=0
+        for i in range (0,numberofseqs):
+            offset=0
+
+            seqitem=items[i].split()
+            seqid=seqitem[0]
+            seq=seqid.strip()+" "*(maxlen-len(seqid)+1)+seqitem[1]
+
+            for k in range(0,numberofblocks-1):
+                offset=offset+numberofseqs
+                item2=items[i+offset].split()
+                seq=seq+item2[1]
+
+            writes.write(seq+"\n")
+        tempitem=seq.split()
+        writes.write('#=GC SS_cons'+" "*(maxlen-len('#=GC SS_cons')+1)+"."*len(tempitem[1])+"\n"+"//")
+        writest=str(writes)
+        writeitem=writest.split("'")
+        writes.close()
+        return str(writeitem[1])
+
+    except Exception as err:
+        exc_type, exc_value, exc_tb = sys.exc_info()
+        tbe = tb.TracebackException(
+            exc_type, exc_value, exc_tb,
+        )
+        log.error(logid+''.join(tbe.format()))
+
+
+### CAVH functions
+
+def count_repetitions(longseq,pattern):
+    results=0
+    len_pattern = len(pattern)
+    for i in range(len(longseq)):
+        if longseq[i:i+len_pattern] == pattern:
+            results += 1
+            print(longseq[i:i+len_pattern],i,i+len_pattern)
+    return results
+
+def search_pattern(longseq, pattern):
+    start = longseq.find(pattern)
+    end = int((start + len(pattern)) -1)
+    return (start, end)
+
+
+def generate_array(longseq, pattern, number, mode):
+    pattern_len = len(pattern)
+    array = []
+    short_sequence = None
+    update_index = None
+    for j in range(0, number):  # (0, n-1)
+        if j == 0:  # Search on the complete sequence
+            start = longseq.find(pattern)
+            end = int((start + pattern_len) - 1)
+            assembly_data = [start, end, mode]
+            array.append(assembly_data)
+            short_sequence = longseq[end + 1:]
+            update_index = longseq.find(short_sequence)
+        else:
+            start_temp = short_sequence.find(pattern)
+            if start_temp != -1:
+                end_temp = int((start_temp + pattern_len) - 1)
+                start = int(start_temp + update_index)
+                end = int(end_temp + update_index)
+                assembly_data = [start, end, mode]
+                array.append(assembly_data)
+                short_sequence = short_sequence[end_temp + 1:]
+                update_index = longseq.find(short_sequence)
+    return array
+
+
+def detect_overlapp_in_array(complete_array):
+    columns = len(complete_array)
+    for i in range(0, columns -1):
+        if complete_array[i+1][0] in range(complete_array[i][0],
+           complete_array[i][1] + 1):
+            return 1
+    return 0
+
+
+def detect_pairs(complete_array, longseq, mat1seq, curmatseq, curmatstar):
+    columns = len(complete_array)  # Complete number columns
+    # Detect overlapping
+    overlapping = detect_overlapp_in_array(complete_array)
+    pair = defaultdict(dict)
+    number_pair = 0
+    if overlapping == 0:
+        for i in range(0, columns-1):  # In array =columns -1,range not reach end#
+            if complete_array[i][-1] != complete_array[i+1][-1]:
+                pair[number_pair] = [complete_array[i], complete_array[i+1]]
+                number_pair = number_pair + 1
+    else:  # At least one pair overlapped with another one
+        substring = None
+        update_index = None
+        for j in range(0, columns -1):  # (0, n-1)
+            if j == 0:  # Search on the complete sequence
+                identity_first = complete_array[j][-1]  # M or S
+                if identity_first == "M":
+                    start_m = longseq.find(curmatseq)  # Start M
+                    end_m = (start_m + len(curmatseq)) - 1  # End M
+                    substring = longseq[end_m + 1:]  # Restrict seq no M
+                    start_temp_s = substring.find(curmatstar)  # Detect start S
+                    update_index = longseq.find(substring)  # Get index
+                    start_s = start_temp_s + update_index  # Update index S
+                    end_s = (start_s + len(curmatstar)) -1
+                    element1 = [start_m, end_m, "M"]
+                    element2 = [start_s, end_s, "S"]
+                    pair[number_pair] = [element1, element2]
+                    number_pair = number_pair + 1
+                else:  # When S is first, search M
+                    start_s = longseq.find(curmatstar)  # Start S
+                    end_s = (start_s + len(curmatstar))-1  # End S
+                    substring = longseq[end_s + 1:]  # Restrict seq no S
+                    update_index = longseq.find(substring)
+                    start_temp_m = substring.find(curmatseq)  # Search M
+                    start_m = start_temp_m + update_index  # Start M
+                    end_m = (start_m + len(curmatseq)) - 1  # End M
+                    element1 = [start_s, end_s, "S"]
+                    element2 = [start_m, end_m, "M"]
+                    pair[number_pair] = [element1, element2]
+                    number_pair = number_pair + 1
+            else:  # This is not the first element
+                if complete_array[j][0] in range(complete_array[j-1][0],
+                                                 complete_array[j-1][1]):
+                    continue
+                else:
+                    identity_first = complete_array[j][-1]  # M or S
+                    if identity_first == "M":
+                        start_temp_m = substring.find(curmatseq)  # Start M in sub
+                        end_temp_m = (start_temp_m + len(curmatseq)) - 1
+                        start_m = start_temp_m + update_index  # Updated M Start
+                        end_m = (start_m + len(curmatseq)) - 1  # Updated M end
+                        substring = substring[end_temp_m + 1:]   # String no start M
+                        update_index = longseq.find(substring)
+                        start_temp_s = substring.find(curmatstar)
+                        if (start_temp_s == -1):  # Not possible to map S
+                            continue
+                        start_s = start_temp_s + update_index  # Updated start S
+                        end_s = (start_s + len(curmatstar)) -1  # Updated end S
+                        element1 = [start_m, end_m, "M"]
+                        element2 = [start_s, end_s, "S"]
+                        pair[number_pair] = [element1, element2]
+                        number_pair = number_pair + 1
+                    else:  # When S is first, search M
+                        start_temp_s = substring.find(curmatstar)  # Start M in sub
+                        end_temp_s = (start_temp_s + len(curmatstar)) - 1
+                        start_s = start_temp_s + update_index  # Updated S Start
+                        end_s = (start_s + len(curmatstar)) - 1  # Updated S end
+                        substring = substring[end_temp_s + 1:]   # String no start S
+                        update_index = longseq.find(substring)
+                        start_temp_m = substring.find(curmatseq)
+                        start_m = start_temp_m + update_index  # Update start M
+                        if (start_temp_m == -1):  # Not possible to map M
+                            continue
+                        end_m = (start_s + len(curmatseq)) -1  # Updated end M
+                        element1 = [start_s, end_s, "S"]
+                        element2 = [start_m, end_m, "M"]
+                        pair[number_pair] = [element1, element2]
+                        number_pair = number_pair + 1
+    return (pair, number_pair)
+
+
+def define_best_pair(pairs, distances_loop, startprecursor,
+                     endprecursor, precursorlen, flanking, longseq):
+    all_distances_sorted = sorted(distances_loop, key=itemgetter(-1))
+    for i in range(len(all_distances_sorted)):
+        if (all_distances_sorted[i][1][0][0] in range(startprecursor, endprecursor + 1)
+                and all_distances_sorted[i][1][0][1] in range(startprecursor, endprecursor + 1)
+                and all_distances_sorted[i][1][1][0] in range(startprecursor, endprecursor + 1)
+                and all_distances_sorted[i][1][1][1] in range(startprecursor, endprecursor + 1)):
+            coord1 = all_distances_sorted[i][1][0][0]
+            coord2 = all_distances_sorted[i][1][1][0]
+            finalcoord2 = all_distances_sorted[i][1][1][1]
+            startfinalseq = coord1-flanking
+            endfinalseq = finalcoord2+flanking
+            correctfinalseq = longseq[startfinalseq:endfinalseq+1]
+            break
+    return (coord1, coord2, correctfinalseq)
+
+
+def find_positions(longseq, mat1seq, curmatseq, curmatstar, flanking):
+    # Features sequences
+    precursorlen = len(mat1seq)
+    # Locate
+    startprecursor = longseq.find(mat1seq)
+    endprecursor = int((startprecursor + precursorlen)-1)
+    # Search the number of substrings along the long sequence
+    mir = count_repetitions(longseq,curmatseq)
+    mirstar = count_repetitions(longseq,curmatstar)
+    mir_array = generate_array(longseq, curmatseq, mir, "M")
+    mirstar_array = generate_array(longseq, curmatstar, mirstar, "S")
+    # All array
+    all_array = mir_array + mirstar_array
+    all_array_sorted = sorted(all_array, key=itemgetter(0))
+    (pairs, number_pairs) = detect_pairs(all_array_sorted, longseq,
+                                         mat1seq, curmatseq, curmatstar)
+    distances_loop = []
+    for pair in pairs:
+        loopsize = abs(pairs[pair][0][1] - pairs[pair][1][0]) - 1
+        distances_loop.append([pair, pairs[pair], loopsize])
+    (coord1, coord2, correctfinalseq) = define_best_pair(pairs, distances_loop,
+                                                         startprecursor,
+                                                         endprecursor,
+                                                         precursorlen,
+                                                         flanking,
+                                                         longseq)
+    return (coord1, coord2, correctfinalseq)
+
+#Utils
 def removekey(d, key):
     logid = scriptname+'.removekey: '
     try:
@@ -154,7 +544,7 @@ def removekey(d, key):
         tbe = tb.TracebackException(
             exc_type, exc_value, exc_tb,
         )
-        clog.error(logid+''.join(tbe.format()))
+        log.error(logid+''.join(tbe.format()))
 
 def getlowest_list(a, n):
     logid = scriptname+'.getlowest_list: '
@@ -172,7 +562,7 @@ def getlowest_list(a, n):
         tbe = tb.TracebackException(
             exc_type, exc_value, exc_tb,
         )
-        clog.error(logid+''.join(tbe.format()))
+        log.error(logid+''.join(tbe.format()))
 
 def gethighest_list(a, n):
     logid = scriptname+'.gethighest_list: '
@@ -190,7 +580,7 @@ def gethighest_list(a, n):
         tbe = tb.TracebackException(
             exc_type, exc_value, exc_tb,
         )
-        clog.error(logid+''.join(tbe.format()))
+        log.error(logid+''.join(tbe.format()))
 
 def getlowest_dict(a, n):
     logid = scriptname+'.getlowest_dict: '
@@ -208,7 +598,7 @@ def getlowest_dict(a, n):
         tbe = tb.TracebackException(
             exc_type, exc_value, exc_tb,
         )
-        clog.error(logid+''.join(tbe.format()))
+        log.error(logid+''.join(tbe.format()))
 
 def gethighest_dict(a, n):
     logid = scriptname+'.gethighest_dict: '
@@ -226,21 +616,7 @@ def gethighest_dict(a, n):
         tbe = tb.TracebackException(
             exc_type, exc_value, exc_tb,
         )
-        clog.error(logid+''.join(tbe.format()))
-
-def toarray(file, ulim=None):
-    logid = scriptname+'.toarray: '
-    try:
-        if not ulim:
-            ulim = 1
-        x = np.loadtxt(str(file), usecols = (ulim), delimiter = '\t', unpack = True, converters = {ulim: lambda s: convertcol(s.decode("utf-8"))})
-        return x
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        clog.error(logid+''.join(tbe.format()))
+        log.error(logid+''.join(tbe.format()))
 
 def convertcol(entry):
     logid = scriptname+'.convertcol: '
@@ -255,134 +631,8 @@ def convertcol(entry):
         tbe = tb.TracebackException(
             exc_type, exc_value, exc_tb,
         )
-        clog.error(logid+''.join(tbe.format()))
+        log.error(logid+''.join(tbe.format()))
 
-def parse_annotation_bed(bed, annotated=None):
-    logid = scriptname+'.parse_annotation_bed: '
-    anno = defaultdict(list)
-    if os.path.isfile(os.path.abspath(bed)):
-        if '.gz' in bed:
-            f = gzip.open(bed,'rt')
-        else:
-            f = open(bed,'rt')
-    else:
-        f = bed
-    try:
-        for line in f:
-            entries = line.rstrip().split('\t')
-            goi = entries[3]
-            if annotated:
-                start = int(entries[10])
-                end   = int(entries[11])-1
-            else:
-                start = int(entries[1])
-                end   = int(entries[2])-1
-            anno[str(goi)].append('-'.join([str(start),str(end)]))
-        return anno
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        clog.error(logid+''.join(tbe.format()))
-
-def readConstraintsFromBed(bed, linewise=None):
-    logid = scriptname+'.readConstraintsFromBed: '
-    cons = defaultdict(list)
-    try:
-        for line in bed:
-            entries = line.rstrip().split('\t')
-            start = int(entries[1])+1
-            end = entries[2]
-            goi = entries[3]
-            if linewise:
-                cons['lw'].append('-'.join([str(start),str(end)]))
-            else:
-                cons[str(goi)].append('-'.join([str(start),str(end)]))
-        return cons
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        clog.error(logid+''.join(tbe.format()))
-
-def readPairedConstraintsFromBed(bed, linewise=None):
-    logid = scriptname+'.readPairedConstraintsFromBed: '
-    cons = defaultdict(list)
-    try:
-        for line in bed:
-            entries = line.rstrip().split('\t')
-            if len(entries) % 2:
-                raise Exception('Unbalanced paired bed, please make sure the paired bed consists of equal number of fields for both constraint entries')
-            else:
-                second = int((len(entries)/2)+1)
-                clog.debug(logid+'For line '+str(entries)+' start of second bed part is field '+str(second))
-            if int(entries[1]) > -1 and int(entries[second]) > -1:
-                start_one = int(entries[1])+1
-                end_one = entries[2]
-                goi = entries[3]
-                start_two = int(entries[second])+1
-                end_two = int(entries[second+1])
-                if linewise:
-                    cons['lw'].append('-'.join([str(start_one), str(end_one), str(start_two), str(end_two)]))
-                else:
-                    cons[str(goi)].append('-'.join([str(start_one), str(end_one), str(start_two), str(end_two)]))
-        return cons
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        clog.error(logid+''.join(tbe.format()))
-
-def readConstraintsFromCSV(csv, linewise=None):
-    logid = scriptname+'.readConstraintsCSV: '
-    cons = defaultdict(
-        lambda: defaultdict(list)
-    )
-
-    try:
-        for line in csv:
-            entries = split(',',line.rstrip())
-            if linewise:
-                cons['def'].append('-'.join([str(entries[1]),str(entries[2])]))
-            else:
-                cons[entries[3]].append('-'.join([str(entries[1]),str(entries[2])]))
-        return cons
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        clog.error(logid+''.join(tbe.format()))
-
-def readConstraintsFromGeneric(generic, linewise=None):
-    logid = scriptname+'.readConstraintsFromGeneric: '
-    cons = defaultdict(
-        lambda: defaultdict(list)
-    )
-
-    try:
-        for line in csv:
-            entries = re.split(r'[ ,|;"]+', line.rstrip())
-            if len(entries > 2):
-                if linewise:
-                    cons['lw'].append('-'.join([str(entries[1]),str(entries[2])]))
-                else:
-                    cons[entries[0]].append('-'.join([str(entries[1]),str(entries[2])]))
-            else:
-                if linewise:
-                    cons['lw'].append('-'.join([str(entries[1]),str(entries[2])]))
-                else:
-                    cons['generic'].append('-'.join(str(entries[1:2])))
-        return cons
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        clog.error(logid+''.join(tbe.format()))
 
 def isvalid(x=None):
     logid = scriptname+'.isvalid: '
@@ -399,7 +649,7 @@ def isvalid(x=None):
         tbe = tb.TracebackException(
             exc_type, exc_value, exc_tb,
         )
-        clog.error(logid+''.join(tbe.format()))
+        log.error(logid+''.join(tbe.format()))
 
 def isinvalid(x=None):
     logid = scriptname+'.isinvalid: '
@@ -416,7 +666,7 @@ def isinvalid(x=None):
         tbe = tb.TracebackException(
             exc_type, exc_value, exc_tb,
         )
-        clog.error(logid+''.join(tbe.format()))
+        log.error(logid+''.join(tbe.format()))
 
 def makeoutdir(outdir):
     logid = scriptname+'.makeoutdir: '
@@ -431,7 +681,7 @@ def makeoutdir(outdir):
         tbe = tb.TracebackException(
             exc_type, exc_value, exc_tb,
         )
-        clog.error(logid+''.join(tbe.format()))
+        log.error(logid+''.join(tbe.format()))
 
 def parseseq(sequence):
     logid = scriptname+'.parseseq: '
@@ -463,276 +713,7 @@ def parseseq(sequence):
         tbe = tb.TracebackException(
             exc_type, exc_value, exc_tb,
         )
-        clog.error(logid+''.join(tbe.format()))
-
-def plot_data(fa, raw, consu, consp, const, xs, cons, saveas, outdir):
-    logid = scriptname+'.plot_data: '
-    anime = []
-    #define xs for constraint line
-    consl = []
-    try:
-        for x in const:
-            consl.append(1.25)
-        width = 16/100*len(fa.seq)
-        height = 9
-        fig = plt.figure(figsize=(width,height),dpi=80)
-        ax1 = fig.add_subplot(111)
-
-        ax2 = ax1.twiny()
-        #   line, = ax.plot([], [], lw=2)
-        plt.title("Blue-- = Unconstraint, Green-. = Unpaired, Red = Paired, Gray = Constraint",y=1.075)
-        ax1.set_ylabel('Prob unpaired')
-        ax1.set_xlabel('Nucleotides')
-        #   plt.xticks(range(0,len(fa.seq)+1),(' '+fa.seq),size='small')
-        #add lines to plot
-        ax1.plot(xs, raw, 'b-', xs, consu, 'g-', xs, consp, 'r-', const, consl, 'k-')
-        ax1.set_xlim(0,len(fa.seq)+1)
-        ax2.set_xlim(ax1.get_xlim())
-        ax1.set_xticks(range(0,len(fa.seq)+1))
-        ax1.set_xticklabels((' '+fa.seq), ha="right")
-        #   ax2.set_xlabel(r"Modified x-axis: $1/(1+X)$")
-        ax2.set_xticks(range(1,len(fa.seq)+1))
-        ax2.set_xticklabels(range(1,len(fa.seq)+1), rotation=45, ha="right")
-        # We change the fontsize of minor ticks label
-        ax1.tick_params(axis='both', which='major', labelsize=8)
-        ax1.tick_params(axis='both', which='minor', labelsize=4)
-        ax2.tick_params(axis='both', which='major', labelsize=5)
-        ax2.tick_params(axis='both', which='minor', labelsize=3)
-        goi, chrom = fa.id.split(':')[::2]
-        strand = str(fa.id.split(':')[3].split('(')[1][0])
-        fig.savefig('StruCons_'+goi+'_'+cons+'.'+saveas)
-        plt.close()
-        #   anime.append(plt.plot(xs, raw, 'b-', xs, consu, 'g-', xs, consp, 'r-', const, consl, 'k-'))
-        #   return anime
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        clog.error(logid+''.join(tbe.format()))
-
-def plot_temp(fa, raw, temp, xs, saveas, outdir):
-    logid = scriptname+'.plot_temp: '
-    try:
-        anime = []
-        #define xs for constraint line
-        width = 16/100*len(fa.seq)
-        height = 9
-        fig = plt.figure(figsize=(width,height),dpi=80)
-        ax1 = fig.add_subplot(111)
-        plt.title("Blue-- = "+temp+" degree",y=1.075)
-        ax1.set_ylabel('Prob unpaired')
-        ax1.set_xlabel('Nucleotides')
-        #add lines to plot
-        ax1.plot(xs, raw, 'b-')
-        ax1.set_xlim(0,len(fa.seq)+1)
-        ax1.set_xticks(range(0,len(fa.seq)+1))
-        ax1.set_xticklabels((' '+fa.seq))
-        # We change the fontsize of minor ticks label
-        ax1.tick_params(axis='both', which='major', labelsize=8)
-        ax1.tick_params(axis='both', which='minor', labelsize=4)
-        goi, chrom = fa.id.split(':')[::2]
-        strand = str(fa.id.split(':')[3].split('(')[1][0])
-        fig.savefig('TempCons_'+goi+'_'+temp+'.'+saveas)
-        plt.close()
-        #   anime.append(plt.plot(xs, raw, 'b-', xs, consu, 'g-', xs, consp, 'r-', const, consl, 'k-'))
-        #   return anime
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        clog.error(logid+''.join(tbe.format()))
-
-def calc_gibbs(fc):
-    logid = scriptname+'.calc_gibbs: '
-    try:
-        return fc.pf()[1]
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-            )
-        clog.error(logid+''.join(tbe.format()))
-
-def get_bppm(tmp, start, end):
-    logid = scriptname+'.get_bppm: '
-    bppm = []
-    try:
-#        clog.debug('\t'.join(map(str,[tmp,start,end])))
-        if start < 0 or end > len(tmp):
-            clog.warning(logid+'start of constraint '+str(start)+' end of constraint '+str(end)+' while length of bpp matrix '+str(len(tmp))+'! Skipping!')
-            return None
-
-        for item in tmp:
-            for i in range(int(start),int(end)+1):
-                if item[i] > 0.0:
-                    bppm.append(str.join('\t',[str(tmp.index(item)), str(i), str(item[i])]))
-        return bppm
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        clog.error(logid+''.join(tbe.format()))
-
-def get_ddg(file):
-    logid = scriptname+'.parseseq: '
-    try:
-        ret = defaultdict()
-        if (isinstance(file, str) and os.path.isfile(file)):
-            if '.gz' in file :
-                res = gzip.open(file,'rt')
-            else:
-                res = open(file,'rt')
-
-            for line in res:
-                clog.debug(logid+line)
-                if 'Condition' in line[0:15]:
-                    continue
-                else:
-                    cond, gibbs, dg, nrg, cons = line.rstrip().split('\t')
-                    if not str(cons) in ret:
-                        ret[str(cons)] = defaultdict()
-                    ret[str(cons)][cond] = float(dg)
-        return ret
-
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        clog.error(logid+''.join(tbe.format()))
-
-def calc_ddg(ddgs):
-    logid = scriptname+'.calc_ddg: '
-
-    try:
-        clog.debug(logid+str(ddgs))
-        ddg = ddgs['constraint_unpaired']+ddgs['secondconstraint_unpaired']-ddgs['bothconstraint_unpaired']-ddgs['unconstraint']
-        """Yi-Hsuan Lin, Ralf Bundschuh, RNA structure generates natural cooperativity between single-stranded RNA binding proteins targeting 5' and 3'UTRs, Nucleic Acids Research, Volume 43, Issue 2, 30 January 2015, Pages 1160-1169, https://doi.org/10.1093/nar/gku1320"""
-
-        return ddg
-
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-            )
-        clog.error(logid+''.join(tbe.format()))
-
-def calc_bpp(bppm):
-    logid = scriptname+'.calc_bpp: '
-    bpp = 0.0;
-    try:
-        for entry in bppm:
-            base, mate, prob = map(float,entry.split('\t'))
-            bpp += prob
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-            )
-        clog.error(logid+''.join(tbe.format()))
-
-    return bpp
-
-def calc_nrg(bpp):
-    logid = scriptname+'.calc_nrg: '
-    #set kT for nrg2prob and vice versa calcs
-    kT = 0.61632077549999997
-    nrg = 0.0;
-    try:
-        if bpp > 0.0:
-            nrg = -1 * kT * math.log(bpp)
-        return nrg
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        clog.error(logid+''.join(tbe.format()))
-
-def print_region_up(data, seqlength=None, region=None):
-    #   pp = pprint.PrettyPrinter(indent=4)#use with pp.pprint(datastructure)
-    #   pp.pprint(data)
-    logid = scriptname+'.print_region_up: '
-    try:
-        if data:
-            ups=''
-            x = int(region)
-            for i in range(int(seqlength)):
-                if isinvalid(data[i][x]):
-                    data[i][x] = np.nan
-                else:
-                    data[i][x] = round(data[i][x],7)
-                ups+=str(i+1)+"\t"+str(data[i][x])+"\n"
-            return ups
-        else:
-            clog.error(logid+'No up data to print')
-            return ups
-
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        clog.error(logid+''.join(tbe.format()))
-
-def print_up(data=None, seqlength=None, region=None):
-    #   pp = pprint.PrettyPrinter(indent=4)#use with pp.pprint(datastructure)
-    #   pp.pprint(data)
-    logid = scriptname+'.print_up: '
-    try:
-        if data:
-            ups=''
-            for i in range(int(seqlength)):
-                for x in range(1,region+1):
-                    if isinvalid(data[i][x]):
-                        data[i][x] = np.nan
-                    else:
-                        data[i][x] = round(data[i][x],7)
-                ups+=str(i+1)+"\t"+"\t".join(map(str,data[i][1:region+1]))+"\n"
-            return ups
-        else:
-            clog.error(logid+'No up data to print')
-            return ups
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        clog.error(logid+''.join(tbe.format()))
-
-def up_to_array(data=None, region=None, seqlength=None):
-    #   pp = pprint.PrettyPrinter(indent=4)#use with pp.pprint(datastructure)
-    #   pp.pprint(data[165553:165588])
-    logid = scriptname+'.up_to_array: '
-    try:
-        if data:
-            entries=[]
-            if not seqlength:
-                seqlength = len(data)
-            if not region:
-                region = slice(1,len(data[0]))
-            for i in range(seqlength):
-                entries.append([])
-                for e in range(len(data[i])):
-                    if isinvalid(data[i][e]):
-                        data[i][e] = np.nan
-                    else:
-                        data[i][e] = round(data[i][e],8)
-                entries[i].extend(data[i][region])
-            return np.array(entries)
-        else:
-            clog.error(logid+'No up data to print')
-            return np.array()
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-        )
-        clog.error(logid+''.join(tbe.format()))
+        log.error(logid+''.join(tbe.format()))
 
 def npprint(a, o=None):#, format_string ='{0:.2f}'):
     logid = scriptname+'.npprint: '
@@ -751,56 +732,7 @@ def npprint(a, o=None):#, format_string ='{0:.2f}'):
         tbe = tb.TracebackException(
         exc_type, exc_value, exc_tb,
         )
-        clog.error(logid+''.join(tbe.format()))
-
-def printdiff(a, o=None):
-    logid = scriptname+'.printdiff: '
-    try:
-        np.save(o, a)
-        #np.savetxt(o, a, delimiter='\t', encoding='bytes')
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-        exc_type, exc_value, exc_tb,
-        )
-        clog.error(logid+''.join(tbe.format()))
-
-def read_precalc_plfold(data, name, seq):
-    logid = scriptname+'.read_precalc_plfold: '
-    try:
-        for i in range(len(seq)):
-            data.append([])
-            data[i] = []
-        with gzip.open(name,'rt') as o:
-            for line in o:
-                cells = line.rstrip().split('\t')
-                data[int(cells[0])-1].append([])
-                data[int(cells[0])-1][0] = None
-                for a in range(1,len(cells)):
-                    data[int(cells[0])-1].append([])
-                    data[int(cells[0])-1][a] = float(cells[a])
-        return data
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-        exc_type, exc_value, exc_tb,
-        )
-        clog.error(logid+''.join(tbe.format()))
-
-def pl_to_array(name, ulim, fmt='npy'):
-    logid = scriptname+'.pl_to_array: '
-    try:
-        clog.debug('\t'.join([logid,name]))
-        if fmt == 'txt':
-            return np.array(np.loadtxt(name, usecols=ulim, unpack=True, delimiter='\t', encoding='bytes'))
-        elif fmt == 'npy':
-            return np.array(np.load(name)[:,ulim])
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-        exc_type, exc_value, exc_tb,
-        )
-        clog.error(logid+' '+name+': '.join(tbe.format()))
+        log.error(logid+''.join(tbe.format()))
 
 def idfromfa(id):
     logid = scriptname+'.idfromfa: '
@@ -809,41 +741,15 @@ def idfromfa(id):
         goi, chrom = id.split(':')[::2]
         strand = str(id.split(':')[3].split('(')[1][0])
     except:
-        clog.error(logid+'Fasta header is not in expected format, you will loose information on strand and chromosome')
+        log.error(logid+'Fasta header is not in expected format, you will loose information on strand and chromosome')
         goi = id
         chrom, strand = ['na','na']
 
     if goi and chrom and strand:
         return [str(goi), str(chrom), str(strand)]
     else:
-        clog.error(logid+'Could not assign any value from fasta header, please check your fasta files')
+        log.error(logid+'Could not assign any value from fasta header, please check your fasta files')
         sys.exit('Could not assign any value from fasta header, please check your fasta files')
-
-def constrain_paired(fc, start, end):
-    logid = scriptname+'.constrain_paired: '
-    try:
-        for x in range(start, end+1):
-            fc.hc_add_bp_nonspecific(x,0) #0 means without direction  ( $ d < 0 $: pairs upstream, $ d > 0 $: pairs downstream, $ d == 0 $: no direction)
-        return fc
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-            )
-        clog.error(logid+''.join(tbe.format()))
-
-def constrain_unpaired(fc, start, end):
-    logid = scriptname+'.constrain_unpaired: '
-    try:
-        for x in range(start, end+1):
-            fc.hc_add_up(x)
-        return fc
-    except Exception as err:
-        exc_type, exc_value, exc_tb = sys.exc_info()
-        tbe = tb.TracebackException(
-            exc_type, exc_value, exc_tb,
-            )
-        clog.error(logid+''.join(tbe.format()))
 
 def print_globaldicts():
     logid = scriptname+'.print_globaldicts: '
@@ -855,7 +761,7 @@ def print_globaldicts():
         tbe = tb.TracebackException(
             exc_type, exc_value, exc_tb,
             )
-        clog.error(logid+''.join(tbe.format()))
+        log.error(logid+''.join(tbe.format()))
 
 
 def print_globallists():
@@ -868,7 +774,7 @@ def print_globallists():
         tbe = tb.TracebackException(
             exc_type, exc_value, exc_tb,
             )
-        clog.error(logid+''.join(tbe.format()))
+        log.error(logid+''.join(tbe.format()))
 
 #def bpp_callback(v, v_size, i, maxsize, what, data):
 #   if what & RNA.PROBS_WINDOW_BPP:
@@ -878,6 +784,5 @@ def print_globallists():
 #   if what & RNA.PROBS_WINDOW_UP:
 #       #    data['up'].extend([{ 'i': i, 'up': v}])
 #       data['up'].extend([v])
-
 #
 # Collection.py ends here
